requirements.txt
beautifulsoup4>=4.8.1
bs4>=0.0.1
google-api-core>=1.14.3
google-auth>=1.7.1
google-cloud-bigquery>=1.22.0
google-cloud-core>=1.0.3
google-resumable-media>=0.5.0
googleapis-common-protos>=1.6.0
python-dateutil>=2.8.0
requests>=2.22.0
soupsieve>=1.9.5
urllib3>=1.25.7
twilio>=6.33.1